move n_effective from utils to electronegativity
